Built-in hash function
This hash method uses python's built-in hash() method. I chose this method because it seems to be the simplest method, and a function that's specifically designed for hashing seemed like it would be good for a hash table.

When collisions happen, I just append the data item to the list stored in the hash table at that index.

Statistics for title hash:
Collisions: 8227
Wasted space: 3227
Construction time: 0.009 seconds

Statistics for quote hash:
Collisions: 7243
Wasted space: 2243
Construction time: 0.007 seconds

This method has approximately the same number of collisions and wasted space as the previous method, but the construction time is almost 10x faster.