squared ascii values method
This hash method takes a list of the ascii values for the characters in the input string, squares them, then uses the sum as the key. I chose this method because I thought it would be random enough but run much faster than the previous methods.

When collisions happen, I just append the data item to the list stored in the hash table at that index.

Statistics for title hash:
Collisions: 8269
Wasted space: 3269
Construction time: 0.021 seconds

Statistics for quote hash:
Collisions: 7217
Wasted space: 2217
Construction time: 0.027 seconds

This ended up being slower than the built-in hash() function.