Linear Probing
For this method, I use the original random seed strategy, but I use Linear Probing for collision handling.
When collisions happen, I search for the next available slot in the table to store the value.

Statistics for title hash:
Collisions: 10587
Wasted space: 0
Construction time: 2.290 seconds

Statistics for quote hash:
Collisions: 10031
Wasted space: 0
Construction time: 2.338 seconds

This method used the most memory, since the size of the hash table had to be less than or equal to the amount of data (in this case 15000).
It also had the most collisions, and the longest construction time by far. However, it had zero wasted space since the table's size was equal to the number of elements stored.